requirements.txt
PyPDF2==3.0.1
pypdf2
python-docx
pdf2image
pdfplumber
pillow
pytesseract
google-generativeai==0.8.0
grpcio==1.65.5
grpcio-status==1.65.5
protobuf==5.27.3
requests
